dashboard track application details issue fix (query correction for pendency day)

+++ src/main/java/in/gov/abdm/nmr/dto/HealthProfessionalApplicationTo.java
@@ -20,7 +20,7 @@ public class HealthProfessionalApplicationTo {
     private String applicantFullName;
     private BigInteger applicationTypeId;
     private String applicationTypeName;
-    private Double pendency;
+    private int pendency;
     private BigInteger workFlowStatusId;
     private String doctorActionDate;
     private String smcActionDate;
